rename util package to config
