+++ src/config/config.go
@@ -1,4 +1,4 @@
-package util
+package config
 
 import (
 	"encoding/json"
