+++ src/config/config_test.go
@@ -1,4 +1,4 @@
-package util
+package config
 
 import (
 	"testing"
