+++ src/manager/swancontext/swancontext.go
@@ -1,13 +1,13 @@
 package swancontext
 
 import (
+	"github.com/Dataman-Cloud/swan/src/config"
 	"github.com/Dataman-Cloud/swan/src/manager/event"
 	"github.com/Dataman-Cloud/swan/src/manager/store"
-	"github.com/Dataman-Cloud/swan/src/util"
 )
 
 type SwanContext struct {
 	Store    store.Store
-	Config   util.SwanConfig
+	Config   config.SwanConfig
 	EventBus *event.EventBus
 }
